move login logic to userservice
