+++ packages/relay/src/routes/signup.ts
@@ -1,56 +1,9 @@
-import { SignupProof } from '@unirep/circuits'
 import { Express } from 'express'
 import { DB } from 'anondb/node'
-import { APP_ADDRESS } from '../config'
 import TransactionManager from '../singletons/TransactionManager'
-import { SnarkProof } from '@unirep/utils'
 import { UserRegisterStatus } from '../enums/userRegisterStatus'
 import { UnirepSocialSynchronizer } from '../synchornizer'
-
-async function signup(
-    publicSignals: string[],
-    proof: SnarkProof,
-    hashUserId: String,
-    fromServer: boolean,
-    synchronizer: UnirepSocialSynchronizer
-) {
-    const signupProof = new SignupProof(
-        publicSignals,
-        proof,
-        synchronizer.prover
-    )
-    const valid = await signupProof.verify()
-    if (!valid) {
-        throw new Error('Invalid proof')
-    }
-    const currentEpoch = synchronizer.calcCurrentEpoch()
-    if (currentEpoch !== Number(signupProof.epoch)) {
-        throw new Error('Wrong epoch')
-    }
-    const appContract = TransactionManager.appContract!!
-    const calldata = appContract.interface.encodeFunctionData('userSignUp', [
-        signupProof.publicSignals,
-        signupProof.proof,
-        hashUserId,
-        fromServer,
-    ])
-
-    // TODO: fix transction twice bug
-    /*
-    const hash = await TransactionManager.queueTransaction(
-        APP_ADDRESS,
-        calldata
-    )
-    */
-
-    const parsedLogs = await TransactionManager.executeTransaction(
-        appContract,
-        APP_ADDRESS,
-        calldata
-    )
-
-    return '' //hash
-}
+import { userService } from '../services/UserService'
 
 export default (
     app: Express,
@@ -80,7 +33,7 @@ export default (
     app.post('/api/signup', async (req, res) => {
         try {
             const { publicSignals, proof, hashUserId, fromServer } = req.body
-            const hash = await signup(
+            const hash = await userService.signup(
                 publicSignals,
                 proof,
                 hashUserId,
